Add an exception filter, cleaned up `handleException`
This commit provides the ability for a filter function to be defined
from outside the library. As a convenience, you can use the
`SentryConfigure` class to provide an array of classes instead.

Cleaned up `handleException` method, making use of try/finally to ensure
the exception is passed to CakePHP after reporting with Sentry.