Allow any Throwable class into the filter
Fixes a bug where what was called the ExceptionFilter would only accept
classes derived from the Exception class - this was not correct as
Error-derived classes will also enter this codepath.

Research found that the root parent class for both Exception and Error
is Throwable, and such the filter is now accepting any Throwable class.